Restructure GAC code
